Update from Azgaar master

+++ modules/ui/world-configurator.js
@@ -48,7 +48,7 @@ function editWorld() {
     elevateLakes();
     const heights = new Uint8Array(pack.cells.h);
     Rivers.generate();
-    pack.cells.h = new Uint8Array(heights);
+    pack.cells.h = new Float32Array(heights);
     defineBiomes();
 
     if (layerIsOn("toggleTemp")) drawTemp();
